Refactor into packages
Divide into more sub-packages: duke.io and duke.data
