+++ src/main/java/duke/data/Storage.java
@@ -1,9 +1,8 @@
-package duke;
+package duke.data;
 
 
 import java.io.BufferedWriter;
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.time.LocalDate;
